add `start` script that brings up the stack without running tests
Splits the integration-test-env orchestrator into a reusable bring-up
phase (`lifecycle.ts`) shared by two entrypoints:

- `pnpm start` — bring up the full stack (incl. seed) and block until
  Ctrl+C, so `pnpm test:integration` can run from another terminal.
- `pnpm start:ci` — renamed from `start`; bring up + run tests + tear
  down. CI flow is unchanged via the `test:integration:ci` root alias.
