+++ package.json
@@ -8,7 +8,7 @@
     "lint:ci": "biome ci",
     "test": "vitest --silent passed-only",
     "test:integration": "vitest --config vitest.integration.config.ts --silent passed-only",
-    "test:integration:ci": "pnpm -F @ensnode/integration-test-env start",
+    "test:integration:ci": "pnpm -F @ensnode/integration-test-env start:ci",
     "audit:osv": "osv-scanner scan source --lockfile pnpm-lock.yaml",
     "typecheck": "pnpm -r --parallel --aggregate-output typecheck",
     "changeset": "changeset",
